+++ .babelrc
@@ -1,3 +1,6 @@
 {
-  "presets": ["@babel/preset-react"]
+  "presets": [
+    "@babel/preset-env",
+    "@babel/preset-react"
+  ]
 }
