+++ package.json
@@ -7,16 +7,21 @@
     "start": "webpack-dev-server",
     "build": "webpack --mode development",
     "eslint": "eslint src/**/*.js src/**/*.jsx",
-    "test": "npm run eslint"
+    "test": "npm run eslint && jest"
   },
   "devDependencies": {
     "@babel/core": "7.4.4",
     "@babel/preset-env": "7.4.4",
     "@babel/preset-react": "7.0.0",
+    "babel-jest": "24.7.1",
     "babel-loader": "8.0.5",
+    "enzyme": "3.9.0",
+    "enzyme-adapter-react-16": "1.12.1",
     "eslint": "5.13.0",
     "eslint-config-htmlacademy": "0.5.2",
     "eslint-plugin-react": "7.12.4",
+    "jest": "24.7.1",
+    "react-test-renderer": "16.8.6",
     "webpack": "4.30.0",
     "webpack-cli": "3.3.1",
     "webpack-dev-server": "3.3.1"
