+++ src/components/card-place/card-place.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import CardPlace from './card-place.jsx';
+import renderer from 'react-test-renderer';
+
+it(`renders is correctly`, () => {
+  const card = renderer
+    .create(<CardPlace/>)
+    .toJSON();
+  expect(card).toMatchSnapshot();
+});
